requirements.txt
apache-airflow==2.8.1
apache-airflow-providers-postgres
pandas
sqlalchemy
psycopg2-binary
python-dotenv